switching to connexion

+++ api/__init__.py
@@ -34,7 +34,6 @@ from gunicorn.app.base import BaseApplication
 
 from api.app import AccountRestService
 from api.config import Config
-from api.middleware import CrossDomain, JSONTranslator
 
 
 logger = logging.getLogger()
@@ -71,20 +70,28 @@ def check_pid(pid_file):
 @click.option("-c", "--config-file", default="/etc/accounting-rest/api.confg", help="Specify configuration file path (creates if not exists)")
 @click.option("--spew/--no-spew", default=False, help="spew all messages, really noisy (default --no-spew)")
 @click.option("--https/--no-https", default=True, help="disable https (default --https)")
+@click.option("--auth/--no-auth", default=True, help="disable authentication (default --auth)")
+@click.option("--direct", is_flag=True, help="don't use gunicorn wrapper")
 @click.pass_context
 @click_log.simple_verbosity_option(logger)
-def cli(ctx, config_file, spew, https):
+def cli(ctx, config_file, spew, https, auth, direct):
     """
     This CLI allows you to manage the Accounting Rest API, this service will run in the background.
     The service is wrapped by gunicorn, and these commands allow you to control the gunicorn master process.
     Check your config file for settings, the default location is in your home folder under `.accounting-rest`.
     This service enables HTTPS by default, so stick it behind a proxy (we recommend nginx).
     Use the --no-https flag only for testing, never in production!
+    :param ctx: our context
+    :param config_file: our configuration file
+    :param spew: very noisy debugging
+    :param https: enable/disable https
+    :param auth: enable/disable authentication mechanism
+    :param direct: don't user Gunicorn
     """
     ctx.obj = {}
 
     def set_level(level):
-        for l in "sqlalchemy", "falcon", "gunicorn", "meinheld", "python-ldap":
+        for l in "sqlalchemy", "connexion", "gunicorn", "meinheld", "python-ldap":
             li = logging.getLogger(l)
             if li:
                 li.setLevel(level)
@@ -104,10 +111,11 @@ def cli(ctx, config_file, spew, https):
     config_file = Config(config_file)
 
     ctx.obj["CONFIG"] = config_file
-    ctx.obj["SRV"] = AccountRestService(config_file, https_only=https)
+    ctx.obj["SRV"] = AccountRestService(config_file, https=https, auth=auth, direct=direct)
     gu_config = config_file.gunicorn()
     gu_config["spew"] = spew
-    ctx.obj["APP"] = GunicornApp(ctx.obj["SRV"], gu_config)
+    if not direct:
+        ctx.obj["APP"] = GunicornApp(ctx.obj["SRV"], gu_config)
 
 
 @cli.command(help="start accounting-rest api")
@@ -118,7 +126,10 @@ def start(ctx):
         logger.info("rest service already running")
     else:
         logger.info("starting rest service")
-        ctx.obj["APP"].run()
+        if ctx.obj["APP"]:
+            ctx.obj["APP"].run()
+        else:
+            ctx.obj["SRV"].start()
 
 
 @cli.command(help="rest service information")
@@ -149,14 +160,21 @@ def info(ctx):
             logger.info("- {0}".format(connection))
         logger.info("*" * 33)
     else:
-        logger.warning("rest service not running, could not find pid")
+        if ctx.obj["APP"]:
+            logger.warning("rest service not running, could not find pid")
+        else:
+            logger.info("started in direct mode, can't determine pid")
 
 
 @cli.command(help="reload rest service")
 @click.pass_context
 def reload(ctx):
     logger.info("reloading rest service")
-    ctx.obj["APP"].reload()
+    if ctx.obj["APP"]:
+        ctx.obj["APP"].reload()
+    else:
+        ctx.obj["SRV"].stop()
+        ctx.obj["SRV"].start()
     logger.info("reload finished")
 
 
@@ -176,25 +194,33 @@ def stop(ctx, force):
         except OSError as e:
             logger.error("failed to stop rest service: {0}".format(e))
     else:
-        logger.warning("could not find pid")
+        if ctx.obj["APP"]:
+            logger.warning("could not find pid")
+        else:
+            ctx.obj["SRV"].stop()
 
 
 @cli.command(help="restart rest service")
 @click.pass_context
 def restart(ctx):
-    pid = check_pid(ctx.obj["CONFIG"].gunicorn().get("pidfile"))
-    if pid:
-        logger.info("stopping rest service")
-        try:
-            os.kill(pid, signal.SIGTERM)
-            time.sleep(3)
-            logger.info("stop finished, restarting rest service")
+    if ctx.obj["APP"]:
+        pid = check_pid(ctx.obj["CONFIG"].gunicorn().get("pidfile"))
+        if pid:
+            logger.info("stopping rest service")
+            try:
+                os.kill(pid, signal.SIGTERM)
+                time.sleep(3)
+                logger.info("stop finished, restarting rest service")
+                ctx.obj["APP"].run()
+            except OSError as e:
+                logger.error("failed to restart rest service: {0}".format(e))
+        else:
+            logging.info("rest service not running, starting it.")
             ctx.obj["APP"].run()
-        except OSError as e:
-            logger.error("failed to restart rest service: {0}".format(e))
     else:
-        logging.info("rest service not running, starting it.")
-        ctx.obj["APP"].run()
+        ctx.obj["SRV"].stop()
+        ctx.obj["SRV"].start()
+        logger.info("restart finished")
 
 
 if __name__ == "__main__":
